Add TimeServiceRuleConfiguration
